pos-store/backend/requirements.txt
appdirs==1.4.4
argcomplete==3.5.2
blinker==1.9.0
bson==0.5.10
click==8.1.7
dnspython==1.16.0
Flask==3.1.0
Flask-PyMongo==2.3.0
importlib_resources==6.4.5
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==3.0.2
pillow==11.0.0
PyJWT==2.10.1
pymongo==3.12.0
python-barcode==0.15.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-escpos==3.1
PyYAML==6.0.2
qrcode==8.0
setuptools==75.6.0
six==1.17.0
Werkzeug==3.1.3
